Deduplicate and separate piano concerns
